deprecate ParseCloud for ParseCloudable
